LayoutTests/imported/w3c/web-platform-tests/navigation-timing/response-start-after-coop-bcg-switch.https-expected.txt


FAIL Navigation responseStart is not clamped to zero after a cross-origin COOP browsing context group switch assert_greater_than: responseStart should not be clamped to zero after the COOP browsing context group switch expected a number greater than 0 but got 0
